Consolidate managed-thread read models
